packages/aegis-auth/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aegis-auth"
version = "0.1.0"
description = "AEGIS Authentication - OIDC, PBAC, multi-tenancy"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pydantic>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.25.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
cognito = ["boto3>=1.34.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/aegis_auth"]
